Pull out test helpers for use in core
